refactor(examples): pick transport via CLI arg, collapse to 2 binaries
Four near-identical executables differed by one line — the TransportConfig
factory. Collapse to one `talker` + one `listener`, each taking the
transport name as the first argument:

  swift run talker   zenoh [tcp/<host>:7447]
  swift run talker   dds   [domain_id]
  swift run listener zenoh [tcp/<host>:7447]
  swift run listener dds   [domain_id]

The second argument still routes to `.zenoh(locator:)` or
`.ddsMulticast(domainId:)` exactly as before, with the same defaults.
Unknown transport names exit with status 2 and a usage hint on stderr.

Tutorial in Sources/Examples/README.md is updated end to end.
