+++ Sources/Examples/README.md
@@ -1,22 +1,31 @@
 # Examples
 
-Four minimal executables that mirror [`demo_nodes_cpp`](https://github.com/ros2/demos/tree/rolling/demo_nodes_cpp)'s `talker` / `listener` — one pair for each transport. Each is ~25 lines of Swift using only the `SwiftROS2` umbrella import.
+Two minimal executables that mirror [`demo_nodes_cpp`](https://github.com/ros2/demos/tree/rolling/demo_nodes_cpp)'s `talker` / `listener`. The transport is picked by the first CLI argument, so one binary covers both Zenoh and DDS.
 
-| Target           | What it does                                      | Transport  |
-|------------------|---------------------------------------------------|------------|
-| `talker_zenoh`   | Publishes `std_msgs/String` on `/chatter` at 1 Hz | Zenoh      |
-| `talker_dds`     | Same, but over DDS                                | CycloneDDS |
-| `listener_zenoh` | Subscribes to `/chatter` and prints each message  | Zenoh      |
-| `listener_dds`   | Same, but over DDS                                | CycloneDDS |
+| Target     | What it does                                      |
+|------------|---------------------------------------------------|
+| `talker`   | Publishes `std_msgs/String` on `/chatter` at 1 Hz |
+| `listener` | Subscribes to `/chatter` and prints each message  |
 
 Message type is `std_msgs/msg/String`, payload is `"Hello World: N"`. Default QoS is `.sensorData` (best-effort, keep-last-10).
+
+## Invocation
+
+```bash
+swift run talker    zenoh [tcp/<host>:7447]   # default locator: tcp/127.0.0.1:7447
+swift run talker    dds   [domain_id]         # default domain_id: 0
+swift run listener  zenoh [tcp/<host>:7447]
+swift run listener  dds   [domain_id]
+```
+
+The first argument selects the transport (`zenoh` or `dds`). The second is transport-specific: a Zenoh router locator or a ROS 2 domain ID. Both arguments default, so `swift run talker` alone targets a local Zenoh router at `tcp/127.0.0.1:7447`.
 
 ## Prerequisites
 
 - macOS with Xcode 16+ **or** Ubuntu 22.04 / 24.04 with Swift 5.9+ and `ros-<distro>-cyclonedds` installed. See the top-level [`README.md`](../../README.md#installation) for per-platform setup.
-- A ROS 2 install on the peer side (Humble / Jazzy / Kilted / Rolling). These demos default to the Jazzy wire format; pass `.distro:` to `ROS2Context` if you need Humble.
+- A ROS 2 install on the peer side (Humble / Jazzy / Kilted / Rolling). These demos default to the Jazzy wire format; edit `.distro:` on the `ROS2Context` call if you need Humble.
 - **Zenoh only:** a running `rmw_zenoh_cpp` router (`ros2 run rmw_zenoh_cpp rmw_zenohd`) that both sides can reach over TCP.
-- **DDS only:** a multicast-capable LAN on a shared `ROS_DOMAIN_ID` (default `0`). On Wi-Fi without multicast, use `.ddsUnicast(peers:)` instead of `.ddsMulticast(domainId:)`.
+- **DDS only:** a multicast-capable LAN on a shared `ROS_DOMAIN_ID` (default `0`). On Wi-Fi without multicast, switch to `.ddsUnicast(peers:)` — see below.
 
 ## Zenoh tutorial
 
@@ -35,10 +44,9 @@ ros2 run rmw_zenoh_cpp rmw_zenohd            # listens on tcp/0.0.0.0:7447
 Terminal A (Swift publisher):
 
 ```bash
-swift run talker_zenoh tcp/<router-host>:7447
+swift run talker zenoh tcp/<router-host>:7447
 # Publishing: 'Hello World: 1'
 # Publishing: 'Hello World: 2'
-# ...
 ```
 
 Terminal B (ROS 2 subscriber):
@@ -65,10 +73,9 @@ ros2 run demo_nodes_cpp talker
 Terminal B (Swift subscriber):
 
 ```bash
-swift run listener_zenoh tcp/<router-host>:7447
+swift run listener zenoh tcp/<router-host>:7447
 # Listening on /chatter...
 # I heard: 'Hello World: 1'
-# I heard: 'Hello World: 2'
 ```
 
 ### 4. Swift ↔ Swift
@@ -77,10 +84,10 @@ Runs entirely within swift-ros2, no ROS 2 install needed on either side:
 
 ```bash
 # Terminal A
-swift run talker_zenoh tcp/<router-host>:7447
+swift run talker   zenoh tcp/<router-host>:7447
 
 # Terminal B
-swift run listener_zenoh tcp/<router-host>:7447
+swift run listener zenoh tcp/<router-host>:7447
 ```
 
 You still need a `rmw_zenohd` router in the middle — Zenoh peers rendezvous through it.
@@ -94,7 +101,7 @@ CycloneDDS discovery is peer-to-peer, so there is no router. Just run on the sam
 Terminal A (Swift publisher):
 
 ```bash
-swift run talker_dds 0             # ROS_DOMAIN_ID = 0
+swift run talker dds 0             # ROS_DOMAIN_ID = 0
 ```
 
 Terminal B (ROS 2 subscriber):
@@ -120,16 +127,17 @@ ros2 run demo_nodes_cpp talker
 Terminal B (Swift):
 
 ```bash
-swift run listener_dds 0
+swift run listener dds 0
 ```
 
 ### Wi-Fi (no multicast)
 
-On networks that drop multicast, edit the demo to use `.ddsUnicast(peers:)`:
+On networks that drop multicast, edit the demo's `.ddsMulticast(...)` call to use `.ddsUnicast(peers:)`:
 
 ```swift
-let ctx = try await ROS2Context(
-    transport: .ddsUnicast(peers: [DDSPeer.peer(address: "192.168.1.10", domainId: 0)], domainId: 0)
+transport = .ddsUnicast(
+    peers: [DDSPeer.peer(address: "192.168.1.10", domainId: 0)],
+    domainId: 0
 )
 ```
 
@@ -163,11 +171,11 @@ for await msg in sub.messages { print(msg.data) }
 await ctx.shutdown()
 ```
 
-Swap `.zenoh(...)` ↔ `.ddsMulticast(...)` to switch transports — everything above step 1 is identical. That's the whole point of the umbrella API.
+Swap `.zenoh(...)` ↔ `.ddsMulticast(...)` to switch transports — everything above step 1 is identical. That's the whole point of the umbrella API, and it's why the talker / listener demos collapse into a single binary each.
 
 ## Troubleshooting
 
 - **`ros2 topic echo` prints nothing but `ros2 topic list` shows `/chatter`** — wire format mismatch. Pin `.distro:` on the Swift side to match the ROS 2 distro (e.g. `.humble` for Humble's pre-type-hash wire schema).
 - **Connection refused on Zenoh** — router isn't running, wrong IP, or firewall blocks TCP `7447`.
 - **DDS sees nothing** — wrong `ROS_DOMAIN_ID`, or the network drops multicast. Switch to `.ddsUnicast`.
-- **`swift run` fails to find `talker_zenoh`** — run it from the repo root (`deps/swift-ros2`), not a subdirectory; SPM resolves targets relative to `Package.swift`.
+- **`swift run` fails to find `talker`** — run it from the repo root (`deps/swift-ros2`), not a subdirectory; SPM resolves targets relative to `Package.swift`.
